sql: Moved a few cluster setting defitions to sqlclustersettings
This commits mechanically moved the following two cluster settings into
the sqlclustersettings package:
 - SecondaryTenantZoneConfigsEnabled
 - SecondaryTenantsAllZoneConfigsEnabled

This will moved one helper function into the sqlclustersettings package:
 - RequireSystemTenantOrClusterSetting

This way, they can be accessed from within the DSC.

Release note: None
