+++ pkg/server/BUILD.bazel
@@ -271,6 +271,7 @@ go_library(
         "//pkg/sql/sessiondatapb",
         "//pkg/sql/sessioninit",
         "//pkg/sql/sessionprotectedts",
+        "//pkg/sql/sqlclustersettings",
         "//pkg/sql/sqlinstance",
         "//pkg/sql/sqlinstance/instancestorage",
         "//pkg/sql/sqlliveness",
